Bug 1421197 - use 64-bit for restyle generation counters to avoid overflow issues. r=heycam
It is expected to use 64-bit for all the restyle generation counters, since the
getter methods all return uint64_t type at present. However, we're using uint32_t
for the actual counter variables, which means the potential overflow issue is not
avoided.

In this patch, we use 64-bit for the restyle generation counters, so the overflow
issue can be avoided as expected.

MozReview-Commit-ID: 2y2afIcuwvc

UltraBlame original commit: 660d64624ea968587b752e8669e7b3b895b22a00

+++ layout/base/RestyleManager.h
@@ -38,12 +38,12 @@ public:
 
   
   
-  uint32_t GetRestyleGeneration() const { return mRestyleGeneration; }
+  uint64_t GetRestyleGeneration() const { return mRestyleGeneration; }
   
   
   
   
-  uint32_t GetUndisplayedRestyleGeneration() const {
+  uint64_t GetUndisplayedRestyleGeneration() const {
     return mUndisplayedRestyleGeneration;
   }
 
@@ -263,8 +263,8 @@ protected:
 
 private:
   nsPresContext* mPresContext; 
-  uint32_t mRestyleGeneration;
-  uint32_t mUndisplayedRestyleGeneration;
+  uint64_t mRestyleGeneration;
+  uint64_t mUndisplayedRestyleGeneration;
   uint32_t mHoverGeneration;
